extend interface to allow for user-provided memory allocators
Introduce a LuUserContext structure which hold all relevant function
pointers and user's context as void pointers. This is a generalization
of the PngReadProc and PngWriteProc aleady present in the API. The
user context will be extended to also handle the warning messages later.

Also provide default implementations for malloc() and  free(). Also,
the existing API for luPngRead() and luPngWrite was kept as wrappers
(dunno if this is a good or bad thing - but removing it will be easy).

Furthermore, const correctness is improved. Wrting the image should work
with const pointers. To do so, PngInfoStruct was extended by an const
pointer called "cimg", and all functions which access the image during
writing will use that pointer.