Allow luImageCreate()/luImageRelease()/luImageExtractBufAndRelease to be called with NULL as user context pointer
In that case, the default context is used, to simplify the API. This is
especially useful if the image was created with the simplified APIs
(luPngRead()/luPngReadFile()) without an explicit user context, so that the
resulting image can directly be freed.

This change also updates example.c to actually free the loaded image
before exiting the program.